Wrote a new server abstraction that works nicely.
However there's a lot of boilerplate code and no documentation.
Fixing that now.
