+++ software/CMakeLists.txt
@@ -22,22 +22,18 @@ add_subdirectory(src)
 
 #LISTS OF SOURCE FILES FOR TARGETS##########################################
 set(SOURCE_FILES_SPI_ADC_TEST src/spi/spi.cpp)
-set(SOURCE_FILES_THREADED_ECHO_SERVER src/counter/threaded_counter.cpp
-        src/server/safe_queue.hpp src/server/queue_items.hpp)
 set(SOURCE_FILES_THREADED_ADC_TEST src/adc/threaded_adc_test.cpp
         src/server/safe_queue.hpp src/server/queue_items.hpp)
 ###########################################################################
 
 #LISTS OF BUILD TARGETS####################################################
 add_executable(spi_adc_test ${SOURCE_FILES_SPI_ADC_TEST}) #A simple test to run the adc on a single channel.
-add_executable(threaded_echo_server ${SOURCE_FILES_THREADED_ECHO_SERVER}) #A simple example echo server for dev later.
 add_executable(threaded_adc_test ${SOURCE_FILES_THREADED_ADC_TEST}) #A simple example echo server for dev later.
 ###########################################################################
 
 
 #DEPENDENCIES##############################################################
 add_dependencies(threaded_adc_test adc_lib.a circ_buff.a timestamps.a listener.a)
-add_dependencies(threaded_echo_server listener.a)
 add_dependencies(spi_adc_test adc_lib.a)
 
 ###########################################################################
@@ -45,5 +41,4 @@ add_dependencies(spi_adc_test adc_lib.a)
 #LINKING###################################################################
 target_link_libraries(spi_adc_test bcm2835)
 target_link_libraries(threaded_adc_test bcm2835 adc_lib.a pthread circ_buff.a timestamps.a listener.a)
-target_link_libraries(threaded_echo_server pthread circ_buff.a listener.a)
 ###########################################################################
