+++ software/src/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(util) # Build all utility libraries first.
 add_subdirectory(echo_server) # Build the echo server test thingy.
+add_subdirectory(counter)
 add_subdirectory(adc) # ADC code has minimal dependencies.
