+++ software/src/counter/counter_worker.hpp
@@ -17,9 +17,11 @@
 
 class echo_worker : public worker {
     public:
+        circular_buffer &buff;
         echo_worker(safe_queue<network_queue_item> &my_qn, safe_queue<work_queue_item> &my_qw,
                            circular_buffer &buff)
                 : worker(my_qn, my_qw)
+                , buff(buff)
         {
         };
